@@ -17,8 +17,10 @@
     "moment": "^2.29.1",
     "orbit-db": "^0.26.0",
     "v-clipboard": "^2.2.3",
-    "vue": "^2.6.12",
+    "vue": "^3.2.0",
     "vue-avatar": "^2.3.3",
+    "vue-router": "^4.0.13",
+    "vuex": "^4.1.0",
     "vuex-persistedstate": "^4.0.0-beta.1"
   },
   "devDependencies": {
